WIP: Sample SpringBoot

+++ application/src/main/java/com/example/application/persistence/ProductRepository.java
@@ -5,11 +5,11 @@ import java.util.*;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.transaction.annotation.Transactional;
 
-public interface ProductRepository extends CrudRepository<ProductEntity, Integer> {
+public interface ProductRepository extends CrudRepository<ProductEntity, Long> {
 	@Transactional(readOnly = true)
-	Optional<ProductEntity> findById(int Id);
+	Optional<ProductEntity> findById(long Id);
 
 	// JMA: JPA >=1.7
 	@Transactional(readOnly = false)
-	Long deleteById(int Id);
+	Long deleteById(long Id);
 }
